public/python-api/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Cors
Flask-SocketIO
python-dotenv
pandas
SQLAlchemy
scikit-learn
scipy
numpy
